@@ -3,7 +3,7 @@ class AnimalsController < ApplicationController
   # GET /animals.json
   #
 
-  before_filter :authenticate_diver!
+  before_action :authenticate_diver!
   load_and_authorize_resource
 
   def index
