@@ -1,7 +1,7 @@
 class DiversController < ApplicationController
  
   
-  before_filter :authenticate_diver!
+  before_action :authenticate_diver!
   load_and_authorize_resource
   
   # GET /divers
