@@ -11,11 +11,11 @@
 // GO AFTER THE REQUIRES BELOW.
 //
 //= require jquery
+//= require jquery-ui
 //= require jquery_ujs
 //= require twitter/bootstrap
-//= require jquery.ui.datepicker
+//= require jquery-ui/datepicker
 //= require dataTables/jquery.dataTables
-//= require dataTables/jquery.dataTables.bootstrap
 //= require jquery_nested_form
 //= require rails.validations
 //= require rails.validations.nested_form
